feat: refactor champion data imports and add utility functions for lane and rank handling
